requirements.txt
flask>=3.0
apscheduler>=3.10
python-telegram-bot>=20.0
requests>=2.31
pandas>=2.0
python-dotenv>=1.0
python-dateutil>=2.8
gunicorn>=21.0
